Moving the picture_grow activity
