+++ app/src/main/res/layout/activity_picture_grow.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="fr.sharpbunny.emergencytag.PictureGrowActivity">
+    tools:context="fr.sharpbunny.emergencytag.activities.PictureGrowActivity">
 
     <ImageView
         android:id="@+id/imageView2"
